+++ crates/primitives/block/src/lib.rs
@@ -263,63 +263,70 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_block_tag() {
-        assert_eq!(BlockTag::Latest, BlockTag::from(starknet_core::types::BlockTag::Latest));
-        assert_eq!(BlockTag::Pending, BlockTag::from(starknet_core::types::BlockTag::Pending));
+    fn test_maybe_pending_block_info() {
+        let tx_hashes_pending = vec![Felt::from(1), Felt::from(2)];
+        let pending = DeoxysPendingBlockInfo::new(PendingHeader::default(), tx_hashes_pending.clone());
+        let pending_as_maybe_pending: DeoxysMaybePendingBlockInfo = pending.clone().into();
+        let tx_hashes_not_pending = vec![Felt::from(3), Felt::from(4)];
+        let not_pending = DeoxysBlockInfo::new(Header::default(), tx_hashes_not_pending.clone(), Felt::from(5));
+        let not_pending_as_maybe_pending: DeoxysMaybePendingBlockInfo = not_pending.clone().into();
 
-        assert_eq!(starknet_core::types::BlockTag::Latest, starknet_core::types::BlockTag::from(BlockTag::Latest));
-        assert_eq!(starknet_core::types::BlockTag::Pending, starknet_core::types::BlockTag::from(BlockTag::Pending));
+        assert_eq!(not_pending_as_maybe_pending.as_nonpending(), Some(&not_pending));
+        assert!(pending_as_maybe_pending.as_nonpending().is_none());
+
+        assert_eq!(pending_as_maybe_pending.as_pending(), Some(&pending));
+        assert!(not_pending_as_maybe_pending.as_pending().is_none());
+
+        assert_eq!(pending_as_maybe_pending.as_block_id(), BlockId::Tag(BlockTag::Pending));
+        assert_eq!(not_pending_as_maybe_pending.as_block_id(), BlockId::Number(0));
+
+        assert_eq!(pending_as_maybe_pending.block_n(), None);
+        assert_eq!(not_pending_as_maybe_pending.block_n(), Some(0));
+
+        assert_eq!(pending_as_maybe_pending.tx_hashes(), &tx_hashes_pending);
+        assert_eq!(not_pending_as_maybe_pending.tx_hashes(), &tx_hashes_not_pending);
+
+        assert_eq!(pending_as_maybe_pending.protocol_version(), &StarknetVersion::default());
+        assert_eq!(not_pending_as_maybe_pending.protocol_version(), &StarknetVersion::default());
+
+        let maybe_pending: DeoxysMaybePendingBlock = DeoxysPendingBlock::new_empty(PendingHeader::default()).into();
+        assert!(DeoxysBlock::try_from(maybe_pending.clone()).is_err());
+        assert!(DeoxysPendingBlock::try_from(maybe_pending.clone()).is_ok());
+    }
+
+    #[test]
+    fn test_block_tag() {
+        let tag = BlockTag::Latest;
+        let tag_converted: starknet_core::types::BlockTag = tag.into();
+        assert_eq!(tag_converted, starknet_core::types::BlockTag::Latest);
+        let tag_back: BlockTag = tag_converted.into();
+        assert_eq!(tag_back, tag);
+
+        let tag = BlockTag::Pending;
+        let tag_converted: starknet_core::types::BlockTag = tag.into();
+        assert_eq!(tag_converted, starknet_core::types::BlockTag::Pending);
+        let tag_back: BlockTag = tag_converted.into();
+        assert_eq!(tag_back, tag);
     }
 
     #[test]
     fn test_block_id() {
-        assert_eq!(BlockId::Hash(Felt::ONE), BlockId::from(starknet_core::types::BlockId::Hash(Felt::ONE)));
-        assert_eq!(BlockId::Number(1), BlockId::from(starknet_core::types::BlockId::Number(1)));
-        assert_eq!(
-            BlockId::Tag(BlockTag::Latest),
-            BlockId::from(starknet_core::types::BlockId::Tag(starknet_core::types::BlockTag::Latest))
-        );
+        let hash = Felt::from(1);
+        let hash_converted: starknet_core::types::BlockId = BlockId::Hash(hash).into();
+        assert_eq!(hash_converted, starknet_core::types::BlockId::Hash(hash));
+        let hash_back: BlockId = hash_converted.into();
+        assert_eq!(hash_back, BlockId::Hash(hash));
 
-        assert_eq!(
-            starknet_core::types::BlockId::Hash(Felt::ONE),
-            starknet_core::types::BlockId::from(BlockId::Hash(Felt::ONE))
-        );
-        assert_eq!(starknet_core::types::BlockId::Number(1), starknet_core::types::BlockId::from(BlockId::Number(1)));
-        assert_eq!(
-            starknet_core::types::BlockId::Tag(starknet_core::types::BlockTag::Latest),
-            starknet_core::types::BlockId::from(BlockId::Tag(BlockTag::Latest))
-        );
-    }
+        let number = 1;
+        let number_converted: starknet_core::types::BlockId = BlockId::Number(number).into();
+        assert_eq!(number_converted, starknet_core::types::BlockId::Number(number));
+        let number_back: BlockId = number_converted.into();
+        assert_eq!(number_back, BlockId::Number(number));
 
-    #[test]
-    fn test_block_info_not_pending() {
-        let block_info = DeoxysBlockInfo {
-            header: Default::default(),
-            block_hash: Felt::ONE,
-            tx_hashes: vec![Felt::TWO, Felt::THREE],
-        };
-        let not_pending_block_info = DeoxysMaybePendingBlockInfo::NotPending(block_info.clone());
-
-        assert!(not_pending_block_info.as_pending().is_none());
-        assert_eq!(not_pending_block_info.as_nonpending().unwrap(), &block_info);
-        assert_eq!(not_pending_block_info.as_block_id(), BlockId::Number(block_info.header.block_number));
-        assert_eq!(not_pending_block_info.block_n(), Some(block_info.header.block_number));
-        assert_eq!(not_pending_block_info.block_hash(), Some(block_info.block_hash));
-        assert_eq!(not_pending_block_info.tx_hashes(), &block_info.tx_hashes);
-        assert_eq!(not_pending_block_info.protocol_version(), &block_info.header.protocol_version);
-    }
-
-    #[test]
-    fn test_block_info_pending() {
-        let block_info = DeoxysPendingBlockInfo { header: Default::default(), tx_hashes: vec![Felt::TWO, Felt::THREE] };
-        let pending_block_info = DeoxysMaybePendingBlockInfo::Pending(block_info.clone());
-
-        assert!(pending_block_info.as_nonpending().is_none());
-        assert_eq!(pending_block_info.as_pending().unwrap(), &block_info);
-        assert_eq!(pending_block_info.as_block_id(), BlockId::Tag(BlockTag::Pending));
-        assert!(pending_block_info.block_n().is_none());
-        assert!(pending_block_info.block_hash().is_none());
-        assert_eq!(pending_block_info.tx_hashes(), &block_info.tx_hashes);
-        assert_eq!(pending_block_info.protocol_version(), &block_info.header.protocol_version);
+        let tag = BlockTag::Latest;
+        let tag_converted: starknet_core::types::BlockId = BlockId::Tag(tag).into();
+        assert_eq!(tag_converted, starknet_core::types::BlockId::Tag(starknet_core::types::BlockTag::Latest));
+        let tag_back: BlockId = tag_converted.into();
+        assert_eq!(tag_back, BlockId::Tag(BlockTag::Latest));
     }
 }